Move core logic outside controller
